requirements.txt
fastmcp
pymongo
python-dotenv
